Comparando as senhas na janela cadastrar

+++ src/Registro.java
@@ -8,7 +8,10 @@ import javax.swing.JButton;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
+import javax.swing.JPasswordField;
 import javax.swing.JTextField;
+import java.util.*;
+import java.io.*;
 
 /*
  * 
@@ -26,8 +29,11 @@ public class Registro extends JFrame implements ActionListener {
 	JLabel usuario = new JLabel("Nome: ");
 	JTextField usuarioTexto = new JTextField();
 
-	JLabel senha = new JLabel("Senha: ");
-	JTextField senhaTexto = new JTextField();
+	JLabel senhaLabel = new JLabel("Senha: ");
+	JPasswordField senhaTexto = new JPasswordField(20);
+	
+	JLabel senhaConfirmacao = new JLabel("Senha*: ");
+	JPasswordField senhaConfirmacaoTexto = new JPasswordField(20);
 	
 	File file = new File("USERDATA.txt");
 
@@ -47,24 +53,18 @@ public class Registro extends JFrame implements ActionListener {
 		
 		String filepath = "usuarios.txt";
 		
-		//Inserção do email
-		email.setBounds(20, 40, 60, 30);
-		emailTexto.setBounds(70, 40, 190, 30);
-		
 		//Inserção do usuario
 		usuario.setBounds(20, 80, 60, 30);
 		usuarioTexto.setBounds(70, 80, 190, 30);
 		
 		
 		//Inserçao da senha
-		senha.setBounds(20, 120, 60, 30);
+		senhaLabel.setBounds(20, 120, 60, 30);
 		senhaTexto.setBounds(70, 120, 190, 30);
 		
 		//Confirmar senha
-		//JLabel senhaConfirmacao = new JLabel("Senha: ");
-		//senha.setBounds(20, 120, 60, 30);
-		//JTextField senhaTexto = new JTextField();
-		//senhaTexto.setBounds(70, 120, 190, 30);
+		senhaConfirmacao.setBounds(20, 160, 60, 30);
+		senhaConfirmacaoTexto.setBounds(70, 160, 190, 30);
 		
 		JButton voltar = new JButton("Voltar");
 		voltar.setBounds(10, 10, 70, 20);
@@ -80,41 +80,49 @@ public class Registro extends JFrame implements ActionListener {
 		registro.setLocationRelativeTo(null);
 		registro.setLayout(null);
 		registro.add(cadastrar);
-		registro.add(emailTexto);
-		registro.add(email);
 		registro.add(usuario);
 		registro.add(usuarioTexto);
-		registro.add(senha);
+		registro.add(senhaLabel);
 		registro.add(senhaTexto);
+		registro.add(senhaConfirmacao);
+		registro.add(senhaConfirmacaoTexto);
 		registro.add(voltar);
+		
+		//Salvando senha(HIDDEN)
+		char[] senha = senhaTexto.getPassword();
 	}
 
 	@Override
 	public void actionPerformed(ActionEvent e) {
+		
 		//Botao cadastrar
 		if(e.getActionCommand().equals("Cadastrar")) {
 			
-			// Escrevendo no arquivo
-			try {
-				FileWriter fw = new FileWriter(file);
-				emailTexto.write(fw);
-				usuarioTexto.write(fw);
-				senhaTexto.write(fw);
-				fw.close();
-			} catch (IOException e1) {
-				e1.printStackTrace();
-				System.out.println("ERROR !!!");
+			if(Arrays.equals(senhaTexto.getPassword(), senhaConfirmacaoTexto.getPassword())) {
+				System.out.println("confirmaçao efetuada");
+			
+				// Escrevendo no arquivo
+				try {
+					FileWriter fw = new FileWriter(file);
+					usuarioTexto.write(fw);
+					senhaTexto.write(fw);
+					fw.close();
+				} catch (IOException e1) {
+					e1.printStackTrace();
+					System.out.println("ERROR !!!");
+				}
+				/********************************/
+				//Fechar e abrir nova janela de logIn
+				registro.setVisible(false);
+				
+				Entrada entrada = new Entrada();
+				entrada.abrirJanela();
+				/********************************/
+		}else {
+			System.out.println("Senha incorreta");
 			}
-			
-			
-			/********************************/
-			//Fechar e abrir nova janela de logIn
-			registro.setVisible(false);
-			
-			Entrada entrada = new Entrada();
-			entrada.abrirJanela();
-			/********************************/
 		}
+		
 		//Botao voltar
 		if(e.getActionCommand().equals("Voltar")) {
 			registro.setVisible(false);
